Construct PlayerBullet with f32 position

+++ src/game/bullets_manager.rs
@@ -37,7 +37,7 @@ impl BulletsManager {
 
         if ticks_since_last >= self.autofire_ticks {
             self.last_bullet_tick = self.current_tick;
-            self.bullets.push(PlayerBullet::new(position.0, position.1, self.bullet_speed))
+            self.bullets.push(PlayerBullet::new(position.0 as f32, position.1 as f32, self.bullet_speed))
         }
     }
